extract toc and add test to ensure that slugger counter is exploited (refs #12)
