+++ package.json
@@ -11,7 +11,8 @@
     "build-demo": "npm run build-demo-01 && npm run build-demo-02",
     "build-demo-01": "rm -rf demo/01-default-layout && node bin/markdown-to-html.js -m convert -l default -O demo/01-default-layout samples/01-default-layout",
     "build-demo-02": "rm -rf demo/02-remarkjs && node bin/markdown-to-html.js -m convert -l remarkjs -O demo/02-remarkjs samples/02-remarkjs",
-    "test": "npm run check-style && nyc --reporter=html --reporter=text --reporter=cobertura mocha",
+    "test": "npm run check-style && npm run unit-test",
+    "unit-test": "nyc --reporter=html --reporter=text --reporter=cobertura mocha",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
     "check-style": "prettier --check \"src/**/*.js\" \"test/**/*.js\" \"bin/*.js\"",
     "fix-style": "prettier --write \"src/**/*.js\" \"test/**/*.js\" \"bin/*.js\""
