+++ src/marked/slugify.js
@@ -1,11 +1,14 @@
 const marked = require('marked').marked;
-const slugger = new marked.Slugger();
 
 /**
  * Convert to lower case and replace whitespaces by '-'
+ *
+ * TODO get rid of the function as Slugger (counter consistency for multiple invocation...)
+ *
  * @param {string} text
  */
 function slugify(text) {
+    const slugger = new marked.Slugger();
     return slugger.slug(text);
 }
